requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.2=py37hecd8cb5_1001
backcall=0.2.0=py_0
blas=1.0=mkl
bzip2=1.0.8=h1de35cc_0
ca-certificates=2020.12.8=hecd8cb5_0
cairo=1.14.12=hc4e6be7_4
certifi=2020.12.5=py37hecd8cb5_0
click=7.1.2=py_0
cycler=0.10.0=py37_0
decorator=4.4.2=py_0
ffmpeg=4.0=h01ea3c9_0
fontconfig=2.13.0=h5d5b041_1
freetype=2.10.4=ha233b18_0
gettext=0.19.8.1=hb0f4f8b_2
glib=2.66.1=h9bbe63b_0
graphite2=1.3.14=h38d11af_0
harfbuzz=1.8.8=hb8d4a28_0
hdf5=1.10.2=hfa1e0ec_1
icu=58.2=h0a44026_3
intel-openmp=2019.4=233
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.19.0=py37h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jasper=2.0.14=h636a363_1
jedi=0.18.0=py37hecd8cb5_0
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=he5867d9_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.0=py37hecd8cb5_0
kiwisolver=1.3.0=py37h23ab428_0
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libopencv=3.4.2=h7c891bd_1
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
libvpx=1.7.0=h378b8a2_0
libxml2=2.9.10=h7cdb67c_3
lightgbm=3.1.1=py37h23ab428_0
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.2=h79c402e_3
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py37h181983e_0
mkl=2019.4=233
mkl-service=2.3.0=py37h9ed2024_0
mkl_fft=1.2.0=py37hc64f4ea_0
mkl_random=1.1.1=py37h959d312_0
ncurses=6.2=h0a44026_1
ninja=1.10.2=py37hf7b0b51_0
nltk=3.5=py_0
numpy=1.19.2=py37h456fd55_0
numpy-base=1.19.2=py37hcfb5961_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openssl=1.1.1i=h9ed2024_0
pandas=1.1.5=py37hb2f4e1b_0
parso=0.7.0=py_0
pcre=8.44=hb1e8313_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.0.1=py37h5270095_0
pip=20.3.3=py37hecd8cb5_0
pixman=0.40.0=haf1e3a3_0
prompt-toolkit=3.0.8=py_0
ptyprocess=0.6.0=pyhd3eb1b0_2
py-opencv=3.4.2=py37h7c891bd_1
pygments=2.7.3=pyhd3eb1b0_0
pyparsing=2.4.7=py_0
python=3.7.9=h26836e1_0
python-dateutil=2.8.1=py_0
pytorch=1.7.1=py3.7_0
pytz=2020.5=pyhd3eb1b0_0
pyzmq=20.0.0=py37h23ab428_1
readline=8.0=h1de35cc_0
regex=2020.11.13=py37h9ed2024_0
scikit-learn=0.24.0=pypi_0
scipy=1.5.2=py37h912ce22_0
setuptools=51.0.0=py37hecd8cb5_2
six=1.15.0=py37hecd8cb5_0
sqlite=3.33.0=hffcf06c_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hb0a8c7a_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cpu
tornado=6.1=py37h9ed2024_0
tqdm=4.55.0=pyhd3eb1b0_0
traitlets=5.0.5=py_0
typing_extensions=3.7.4.3=py_0
wcwidth=0.2.5=py_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.3=hb1e8313_3
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
